backend/pyproject.toml
[project]
name = "medichain-backend"
version = "0.1.0"
description = "Decentralized Clinical Trial Matching via AI Agent Mesh & On-Chain Trust"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "MediChain Team", email = "team@medichain.io" }
]
keywords = ["healthcare", "ai", "blockchain", "clinical-trials", "singularitynet"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    # Core Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    # Database
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    # AI/ML
    "google-generativeai>=0.8.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "faiss-cpu>=1.9.0",
    "qdrant-client>=1.12.0",
    "sentence-transformers>=3.3.0",
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=43.0.0",
    "httpx>=0.28.0",
    # Healthcare Data
    "fhir.resources>=7.1.0",
    "hl7>=0.4.5",
    "pypdf>=5.1.0",
    # Blockchain/Web3
    "web3>=7.6.0",
    "eth-account>=0.13.0",
    # SingularityNET Decentralized AI Platform
    "snet-sdk>=1.0.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    # Utilities
    "python-multipart>=0.0.17",
    "aiofiles>=24.1.0",
    "redis>=5.2.0",
    "celery>=5.4.0",
    "structlog>=24.4.0",
    "orjson>=3.10.0",
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.1",
    "email-validator>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = ["E501", "B008", "B904"]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
